+++ pkgs/collection/lib/src/iterable_extensions.dart
@@ -7,7 +7,6 @@ import 'dart:math' show Random;
 import '../collection.dart' show DelegatingIterable;
 import 'algorithms.dart';
 import 'functions.dart' as functions;
-import 'utils.dart';
 
 /// Extensions that apply to all iterables.
 ///
@@ -165,10 +164,17 @@ extension IterableExtension<T> on Iterable<T> {
   /// Creates a sorted list of the elements of the iterable.
   ///
   /// The elements are ordered by the natural ordering of the
-  /// property [keyOf] of the element.
-  List<T> sortedBy<K extends Comparable<K>>(K Function(T element) keyOf) {
+  /// [keyOf] property when [descending] is `false` (the default) or reverse
+  /// ordering when [descending] is `true`.
+  List<T> sortedBy<K extends Comparable<K>>(
+    K Function(T element) keyOf, {
+    bool descending = false,
+  }) {
+    final compare = descending
+        ? (K a, K b) => b.compareTo(a)
+        : (K a, K b) => a.compareTo(b);
     var elements = [...this];
-    mergeSortBy<T, K>(elements, keyOf, compareComparable);
+    mergeSortBy<T, K>(elements, keyOf, compare);
     return elements;
   }