+++ pkgs/crypto/test/sha1_test.dart
@@ -32,7 +32,7 @@ void main() {
     test('close closes the underlying sink', () {
       var inner = ChunkedConversionSink<Digest>.withCallback(
         expectAsync1((accumulated) {
-          expect(accumulated.length, equals(1));
+          expect(accumulated, hasLength(1));
           expect(
             accumulated.first.toString(),
             equals('da39a3ee5e6b4b0d3255bfef95601890afd80709'),